requirements.txt
fastmcp>=0.3.0
agno>=0.3.0
requests>=2.31.0
pydantic>=2.7.0
lief>=0.14.0
capstone>=5.0.1
yara-python>=4.3.1
androguard>=3.4.0
pytest>=8.0.0
rich>=13.7.0
